organise components into folder
